refactor(scripts): move context metadata core into scripts/context/ (P3)
Phase P3 of the scripts/ reorganization (issue 1877). Moves the nine files of
the documentation and instruction metadata core into scripts/context/ and
updates every caller in the same PR. Basenames stay byte-identical, as ADR 0064
requires.

This cluster lands before scripts/docs/ (P4) and scripts/pr/ (P5) because
docs-audit-helpers.mjs, docs-audit.mjs, docs-garden-issue.mjs,
docs-navigation-eval.mjs, docs-navigation-eval-result.mjs and
review-materiality-context.mjs all import from it. Their import specifiers
update here while those files themselves stay flat.

The gh-issue lifecycle also splits out. scripts/lib/gh-issue-lifecycle.mjs now
owns runGh, ghPaginate, assertAuthorizedGardenWorkflow, ensureLabelsExist and
LABEL_DEFINITIONS, plus the marker-block parser, page normalizer and
queue-state arbitration that docs-garden-issue-helpers.mjs and
docs-navigation-eval-helpers.mjs each defined separately. That removes the only
entrypoint-imports-entrypoint edge in scripts/: docs-navigation-eval.mjs
imported four symbols from the docs-garden-issue.mjs entrypoint. Error
messages, decision branches and marker schemas stay per-caller and unchanged;
both docs-garden files stay flat and move in P4.

The shared marker parser returns undefined, not null, for "this issue does not
carry my marker". A marker whose payload parses to literal null would otherwise
read as untracked, hide a live queue issue and let a duplicate through. Both
suites gain an assertion for that case, and both fail when the sentinel is
changed back.

Two path derivations inside the moved tests were depth-dependent and are fixed:
the budget suite and the docs-index registry test each walked one directory up
to reach the repository root and now walk two.

ADR 0040 and ADR 0041 name the new module in their Evidence lists; each records
behaviour that now lives there.

Swept: the five package.json aliases and their verbatim copies in the
pre-install validator, ci.yml, supply-chain.yml, .trunk/trunk.yaml, ADR 0005,
ADR 0064, docs/context-standards.md, three focused arms in
agent-quality-gate.sh, and the matching gate-test assertions. The gate gains an
arm for the new shared module routing both consumers' suites, and the gate test
asserts it. The doc-garden skill mirrors reference only pnpm aliases, which do
not change, so they need no edit.
