+++ scripts/check-agent-quality-gate-package-scripts.sh
@@ -19,11 +19,11 @@ const expectedScripts = {
   "agent:prewarm:test": "node scripts/agent-prewarm.test.mjs",
   "agent:review-materiality": "node scripts/review-materiality.mjs",
   "agent:review-materiality:test": "node scripts/review-materiality.test.mjs",
-  "agent:context-check": "node scripts/check-agent-context.mjs",
-  "agent:context-budget": "node scripts/agent-context-budget.mjs",
-  "agent:context-budget:test": "node scripts/agent-context-budget.test.mjs",
-  "docs:index": "node scripts/docs-index.mjs",
-  "docs:index:test": "node scripts/docs-index.test.mjs",
+  "agent:context-check": "node scripts/context/check-agent-context.mjs",
+  "agent:context-budget": "node scripts/context/agent-context-budget.mjs",
+  "agent:context-budget:test": "node scripts/context/agent-context-budget.test.mjs",
+  "docs:index": "node scripts/context/docs-index.mjs",
+  "docs:index:test": "node scripts/context/docs-index.test.mjs",
   "docs:audit": "node scripts/docs-audit.mjs",
   "docs:audit:test": "node scripts/docs-audit.test.mjs",
   "docs:garden": "node scripts/docs-garden-issue.mjs",
