+++ scripts/context/check-agent-context-helpers.mjs
