+++ scripts/context/check-agent-context.mjs
