+++ scripts/context/claude-runtime-document-registry.mjs
