+++ scripts/context/docs-index-helpers.mjs
